Algumas coisas foram arrumadas

+++ src/main/java/br/edu/ifsp/arq/dw2s6/Achados/repository/CupomRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.ifsp.arq.dw2s6.Achados.domain.model.Cupom;
 import br.edu.ifsp.arq.dw2s6.Achados.domain.model.User;
 
-public interface CupomRepository extends JpaRepository<Cupom, Integer>{
+public interface CupomRepository extends JpaRepository<Cupom, Long>{
 	List<Cupom> findByUser(User user);
 }
